feature/VJ-714 - Accessibility issues on Vendor platform - HAT-9

+++ Beis.VendorManagement.Web/Models/Pricing/MinimumCommitmentViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class MinimumCommitmentViewModel : BaseViewModel
     {
-        public int ProductPriceId { get; set; }
+        public long ProductPriceId { get; set; }
 
         public long ProductId { get; set; }
